ran the code through the eclipse formatter it seems to be doing a decent job
i have started using eclipse so that i can refactor the code
up until now i was using vi, but it was getting too hard to work across multiple files

the aim is to rationalise the code so that it looks like a single body of work. at the moment there is camelcase and _'s mixed in throughout the code. it also needs to be C++'d